requirements.txt
github-app-handler==0.23.0
sentry-sdk==1.43.0
flask==3.0.2
Polidoro-PyGithub==2.2.0
python-dotenv==1.0.1
pymdown-extensions==10.7.1
boto3==1.34.70
pydantic==2.6.4
